20160930_2 Prototype

+++ unity-sample-environment/Assets/Scripts/State.cs
@@ -7,8 +7,8 @@ namespace MLPlayer {
 		public bool endEpisode;
 		public byte[][] image;
 		public byte[][] depth;
-		public byte[][] gene;  // add Naka
-		public byte[] rewards; // add Naka
+		public float[][] gene;  // add Naka
+		public float[] rewards; // add Naka
 		public void Clear() {
 			reward = 0;
 			endEpisode = false;
